+++ src/FieldLayout/Contracts/FieldLayoutFormElementProviderInterface.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace CraftCms\Cms\FieldLayout\Contracts;
 
-use CraftCms\Cms\Cp\FormDefinitions\Elements\FormElement;
+use CraftCms\Cms\Cp\Components\ViewComponent;
+use CraftCms\Cms\Cp\FormDefinitions\Contracts\ProjectableFormElement;
 use CraftCms\Cms\FieldLayout\FieldLayoutFormElementContext;
 
 interface FieldLayoutFormElementProviderInterface
 {
-    public function formElement(FieldLayoutFormElementContext $context): ?FormElement;
+    public function formElement(
+        FieldLayoutFormElementContext $context,
+    ): (ViewComponent&ProjectableFormElement)|null;
 }